Split offline tests
